requirements.txt
fastapi==0.111.1
uvicorn==0.26.1
numpy==1.25.2
Pillow==10.0.0
requests==2.31.0
tensorflow==2.14.0